pyproject.toml
[tool.poetry]
name = "parse-site-jleague"
version = "0.1.0"
description = ""
authors = ["Andrey Kagadi <kagadi.a@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests-html = "^0.10.0"
pyppeteer = "^1.0.2"
gigachat = "^0.1.8"
python-dotenv = "^1.0.0"
pytelegrambotapi = "^4.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
